fix: small fixes to succeed the requirements
